requirements.txt
mwapi
python-dotenv
